fixed bonusText
