+++ src/app/pages/Liquidity/subPages/Withdraw/components/WithdrawInput.tsx
@@ -15,6 +15,7 @@ import { LiquidityPageActions } from "app/pages/Liquidity/slice";
 import { WithdrawButton } from "./withdrawButton";
 import { PercentageText } from "./percentageText";
 import { mobile } from "styles/media";
+import { BonusPercent } from "./bonusPercent";
 import { BonusText } from "./bonusText";
 
 type TParams = { poolIndex: string };
@@ -67,13 +68,13 @@ export const WithdrawInput: FC = () => {
           <Grid container justifyContent="space-between">
             <Grid item>
               <SubtitleText variant="body2">
-                {t(translations.LiquidityPage.Withdraw.Bonus())} (%):
+                <BonusText /> (%):
               </SubtitleText>
             </Grid>
 
             <Grid item>
               <SubtitleText variant="body2">
-                <BonusText />
+                <BonusPercent />
               </SubtitleText>
             </Grid>
           </Grid>
